Messaging and Core Logic * Removed entire PanStateLogic and moved all items into the core file * Basics of messaging system listen
